fix: refactor code and tests
